daemon: move blob `jsonrpc_` methods to an individual module
